requirements.txt
requests
python-dotenv
tuya-connector-python
